// src/app/components/model/cliente.interface.ts
export interface Cliente {
	id_cliente: number;
	nombre1_cliente: string;
	nombre2_cliente: string;
	apellido1_cliente: string;
	apellido2_cliente: string;
	cedula_cliente: string;
	ciudad_cliente: string;
	direccion_cliente: string;
	telefono_cliente: string;
	celular_cliente: string;
	email_cliente: string;
	estado_cliente: string;
}